Switch from oneshot to simple and don't daemonize
Since gnome-keyring-daemon started by PAM will be stopped when the user loggs out,
oneshot + RemainAfterExit will fail to run again when logging in for the second time.
Running the initialization in foreground will make it hang around and instead stop
when the user loggs out, allowing it to start again at the second login.